% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAQpress.R
\name{calcAQpress}
\alias{calcAQpress}
\title{Propagule pressure calculation}
\usage{
calcAQpress(AQsites, rivercoords, inventory, dir)
}
\arguments{
\item{AQsites}{Your aquaculture site dataframe. Must have columns: Site.ID, Lat, Long, prov}

\item{rivercoords}{Your river coordinate dataframe. Must have columns: River, Lat, Long}

\item{inventory}{Your inventory dataframe. Must have columns: Site.ID, Lat, Long, Year, prov, totalfish (* totalfish does not have to be 100% accurate as it will be replaced by 1's to indicate that a site was stocked in a given year. Do not include 0's.)}

\item{dir}{= directory where you would like to write csv files}
}
\description{
calcAQpress will calculate propagule pressure for rivers using information on the locations and use of aquaculture sites.
All coordinates must be in decimal degrees, with E/W hemisphere specified using +/-
Make sure that all of your coordinates are in the water. There is a check in this function that will fail if coordinates are not properly placed.
This means that river coordinates should be assigned to the river's mouth. Using an upstream location will fail, as the raster resolution may place such locations on land.
}

