% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAQpress.R
\docType{data}
\name{AQsites}
\alias{AQsites}
\title{Propagule pressure calculation}
\format{An object of class \code{data.frame} with 3 rows and 4 columns.}
\usage{
AQsites
}
\arguments{
\item{AQsites}{Your aquaculture site dataframe. Must have columns: Site.ID, Lat, Long, prov}

\item{rivercoords}{Your river coordinate dataframe. Must have columns: River, Lat, Long}

\item{inventory}{Your inventory dataframe. Must have columns: Site.ID, Lat, Long, Year, prov, totalfish (Note: totalfish does not have to be perfectly accurate as it will be replaced by 1's to indicate that a site was stocked in a given year. Do not include 0's for unstocked (fallow) sites.)}

\item{dir}{The directory where you would like to write csv files}
}
\description{
calcAQpress will calculate propagule pressure for rivers using information on the locations and use of aquaculture sites.

Propagule pressure is calculated for each river, and is equal to the sum (for all aquaculture sites) of 1/(distance between the river and each stocked aquaculture site).

All coordinates must be in decimal degrees, with E/W hemisphere specified using +/-, and all of your coordinates must be in the water.
When in doubt, assign river coordinates to the river's mouth.
Using an upstream location may cause the function to fail, as the raster resolution may place such locations on land.

Two CSV files are created and saved to your specified working directory. prop.press_final.csv is the year-specific propagule pressure for each river. prop.press_avg.csv is the average propagule pressure for each river.
}
\keyword{datasets}

